requirements.txt
boto
python-keyczar
